Refactored file reading code, fixed filepaths for Link colliders.
