+++ fast64_internal/oot/skeleton/importer/functions.py
@@ -3,7 +3,8 @@ from ....f3d.f3d_gbi import F3D
 from ....f3d.f3d_parser import getImportData, parseF3D
 from ....utility import hexOrDecInt, applyRotation
 from ...oot_f3d_writer import ootReadActorScale
-from ...oot_model_classes import OOTF3DContext, ootGetIncludedAssetData
+from ...oot_model_classes import OOTF3DContext
+from ...file_reading import ootGetIncludedAssetData
 from ...oot_utility import ootGetObjectPath, getOOTScale
 from ...oot_texture_array import ootReadTextureArrays
 from ..constants import ootSkeletonImportDict
